// src/generated/types/CreateCourseAccessDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const createCourseAccessDtoSectionEnum = {
  grades: 'grades',
  events: 'events',
  content: 'content',
  groups: 'groups',
} as const

export type CreateCourseAccessDtoSectionEnum = (typeof createCourseAccessDtoSectionEnum)[keyof typeof createCourseAccessDtoSectionEnum]

export type CreateCourseAccessDto = {
  /**
   * @type string
   */
  userId: string
  /**
   * @type string
   */
  courseId: string
  /**
   * @type string
   */
  section: CreateCourseAccessDtoSectionEnum
  /**
   * @default false
   * @type boolean
   */
  canView: boolean
  /**
   * @default false
   * @type boolean
   */
  canEdit: boolean
  /**
   * @default false
   * @type boolean
   */
  canDelete: boolean
}