// src/generated/types/AssistantListDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { CourseAccessPermissionDto } from './CourseAccessPermissionDto.ts'

export type AssistantListDto = {
  /**
   * @type string
   */
  id: string
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  email: string
  /**
   * @type string
   */
  username: string
  /**
   * @default false
   * @type boolean
   */
  hasAccess: boolean
  /**
   * @type array
   */
  permissions: CourseAccessPermissionDto[]
}